Mass-assignment whitelist awareness.
Uses a block instead of hash-assignment to avoid falling foul of mass
assignment whitelist protection.

Adds attr_accessible to models to disallow mass-assignment.  There's
negligible chance anybody ever writing some code that'll be vulnerable (I
hope), but it's worth having it in there so the tests exercise the block
change! :)